Add files via upload
This is code that creates a grid of numbers that represent each day of the month, which can then be clicked. At the moment, it simply gives an alert, but we can link this to the code that displays what day it is to display the day that is clicked.