Now with working boxes!
Fixed error where the boxes were being drawn to fit the whole window as opposed to just the canvas element.

Also, small recommendation: can we figure out a way to make the c and ctx variables global? It's weird that we have to keep making calls to define them every time we want to use them.